refactor(chain): extract `sendTransaction` into separate module
