+++ src/aens.ts
@@ -12,7 +12,8 @@ import { commitmentHash, isAuctionName } from './tx/builder/helpers';
 import { Tag, AensName, ConsensusProtocolVersion } from './tx/builder/constants';
 import { Encoded, Encoding } from './utils/encoder';
 import { UnsupportedProtocolError } from './utils/errors';
-import { sendTransaction, SendTransactionOptions, getName } from './chain';
+import { getName } from './chain';
+import { sendTransaction, SendTransactionOptions } from './send-transaction';
 import { buildTxAsync, BuildTxOptions } from './tx/builder';
 import { TransformNodeType } from './Node';
 import { NameEntry, NamePointer } from './apis/node';
